Shaders v0.3 - Swamp!
IceShader now has the option to increase the lightness
Minor changes to LeavesShader
Swamp will still require more work, but for now it allows two textures
where one will decide the actual texture, the other the transperency of
the object

+++ MMO/Assets/Shaders/LeavesShader.shader
@@ -41,7 +41,7 @@
  
 			//fragment function
 			fixed4 frag(vertexOutput i) : COLOR {
-				float4 tex = tex2D(_MainTex, i.tex.xy * _MainTex_ST.xy + _MainTex_ST.zw);
+				fixed4 tex = tex2D(_MainTex, i.tex.xy * _MainTex_ST.xy + _MainTex_ST.zw);
 				if (tex.a < _Cutoff) {
 				   discard;
 				}
